addressing things
