+++ barretenberg/cpp/src/barretenberg/vm2/simulation/standalone/hybrid_execution.hpp
@@ -4,6 +4,10 @@
 
 namespace bb::avm2::simulation {
 
+// This class is used in fast simulation only.
+// It overrides the execution loop (to remove overhead) but it uses all the other
+// methods from the "gadget" Execution class. That is, dispatching and the opcodes'
+// implementations are shared.
 class HybridExecution : public Execution {
   public:
     using Execution::Execution;
